Split `ResourceSearchPath` to its own file
