+++ srcStatic/ConsoleModuleLoader.h
@@ -31,19 +31,3 @@ private:
 
 	void LoadModuleDll(Module& moduleInfo);
 };
-
-
-class ResManager;
-
-
-class ResourceSearchPath {
-public:
-	static void Set(std::vector<std::string> paths);
-	static void Activate();
-
-private:
-	friend ResManager;
-	static void HookFileSearchPath();
-	static bool CallOriginalGetFilePath(const char* resourceName, /* [out] */ char* filePath);
-	static std::vector<std::string>& ModuleDirectories();
-};
